pyproject.toml

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "Non configured pipeline"
authors = [
  { name = "Fady Adel" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "dagshub>=0.5.9",
    "dvc>=3.59.2",
    "dvc-s3>=3.2.0",
    "hydra-core>=1.3.2",
    "loguru",
    "mkdocs",
    "mlflow>=2.22.0",
    "pandas>=2.2.3",
    "pip",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.6.1",
    "skore>=0.8.3",
    "tqdm",
    "typer",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true
